+++ server/user.h
@@ -91,7 +91,7 @@ extern int init_thread_queue( struct thread *thread );
 extern int attach_thread_input( struct thread *thread_from, struct thread *thread_to );
 extern void detach_thread_input( struct thread *thread_from );
 extern void post_message( user_handle_t win, unsigned int message,
-                          unsigned int wparam, unsigned int lparam );
+                          unsigned long wparam, unsigned long lparam );
 extern void post_win_event( struct thread *thread, unsigned int event,
                             user_handle_t win, unsigned int object_id,
                             unsigned int child_id, void *proc,
